+++ src/com/company/EmployeeSolution.java
@@ -351,17 +351,6 @@ Output: [1,2,2,3,5,6]
         return max_sum_so_far;
     }
 
-    public void merge(int[] nums1, int m, int[] nums2, int n) {
-        int i = m - 1, j = n - 1, k = m + n - 1;
-        while (j > -1) {
-            if (i > -1) {
-                nums1[k--] = nums1[i] > nums2[j] ? nums1[i--] : nums2[j--];
-            } else {
-                nums1[k--] = nums2[j--];
-            }
-        }
-    }
-
     public boolean splitArraySameAverage(int[] nums) {
         int n = nums.length, total = 0;
         for (int i = 0; i < n; i++) total += nums[i];
@@ -4674,6 +4663,7 @@ Output: [1,2,2,3,5,6]
 
         return s;
     }
+
 
 
 }
