+++ src/com/company/SubSet2.java
@@ -438,27 +438,6 @@ class SubSet2 {
             return (int) (start - 1);
         }
 
-        // O(n+m), O(1)
-	/*
-     This problem is called Merge Sorted Array
-     Example:-
-     Input: nums1 = [1,2,3,0,0,0], m = 3, nums2 = [2,5,6], n = 3
-Output: [1,2,2,3,5,6]
-    Algo :- We start filling the array from right end till all elements of nums1 is consumed
-        After that remaining element of nums2 is utitlised
-        */
-
-        public void merge(int[] nums1, int m, int[] nums2, int n) {
-            int i = m - 1, j = n - 1, k = m + n - 1;
-            while (j > -1) {
-                if (i > -1) {
-                    nums1[k--] = nums1[i] > nums2[j] ? nums1[i--] : nums2[j--];
-                } else {
-                    nums1[k--] = nums2[j--];
-                }
-            }
-        }
-
 
         // This is based on binary search of peak element
         // Input: nums = [1,2,3,1]
