+++ src/com/company/TreeNode.java
@@ -472,29 +472,6 @@ class TreeNode {
         return curr;
     }
 
-    // Write a function to return a prime or not
-    // 4 , 7, 11
-    public boolean isPrime(int n) {
-        for (int i = 3; i < Math.sqrt(n); i += 2) {
-            if (n % i == 0) return false;
-        }
-        return true;
-    }
-
-    public int nthPrimeNumber(int n) {
-        int counter = 0;
-        int number = 2;
-        while (true) {
-            if (isPrime(number)) {
-                counter++;
-            }
-            number++;
-            if (counter == n) break;
-        }
-        return number;
-    }
-
-
     /*
     // Algo :-
      * Iterate through all array elements
@@ -633,4 +610,6 @@ class TreeNode {
         tail.next = curr;
         return head;
     }
+
+
 }
